# Use slices in bilinear resampler