This PR changes the bilinear resampler to return data and mask slices from `get_bil_info()` method instead of the `valid_input_index` and `index_array`. The current version uses these latter to compute the slices every time, so some time will be saved.

Also includes "few" flake8 fixes in `geometry.py` forced by `pre-commit`.

 - [ ] Closes #xxxx <!-- remove if there is no corresponding issue, which should only be the case for minor changes -->
 - [ ] Tests added <!-- for all bug fixes or enhancements -->
 - [ ] Tests passed <!-- for all non-documentation changes -->
 - [x] Passes ``git diff origin/master **/*py | flake8 --diff`` <!-- remove if you did not edit any Python files  -->
 - [ ] Fully documented <!-- remove if this change should not be visible to users, e.g., if it is an internal clean-up, or if this is part of a larger project that will be documented later -->