@pnuu 2019-09-10
Note: this will break bilinear interpolation for Satpy until https://github.com/pytroll/satpy/pull/897 is ready.

I'm not yet sure if, and how, to document this. The default bilinear resampler is the Numpy based version, and having two *almost* identical versions documented might be quite confusing.

Appveyor and Travis fail in a place where I have `dask` arrays locally (in linux), but they have `np.ndarray`, not sure what's happening there. The dask versions are the same in all three, 2.3.0. I'll try something.

And I'm going to ignore the "errors" Stickler reported.