@pnuu 2019-09-10
That seemed to work.

For reference some timings (OMP_NUM_THREADS=1 DASK_NUM_WORKERS=4 PYTROLL_CHUNK_SIZE=1024, 2 composites with HRV data, 6 composites without HRV data, `generate=False`):
**current `pyresample` and `satpy` `master` branches**
- first run: 146.2 s
- cached run: 43.3 s
**PR branches of `pyresample` and `satpy`:
- first run: 134.0 s
- cached run: 27.2 s